move menu to a seperate package
